archive_console/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
# Same yt-dlp stack as monthly drivers (archive_*_run.py, one-off).
yt-dlp[default]>=2024.12.13
# Galleries tab + Supported sites (`gallery-dl.exe` lives next to python.exe under .venv\Scripts).
gallery-dl>=1.26.0
jinja2==3.1.4
pydantic==2.10.3
pytest==8.3.4
httpx==0.28.1
apscheduler>=3.10.4
pystray>=0.19.5
Pillow>=10.0.0
